+++ apps/console/src/extensions/configs/application.tsx
@@ -415,6 +415,7 @@ export const applicationConfig: ApplicationConfig = {
                     applicationRoles?.enabled) 
                 && (
                     application?.templateId === ApplicationManagementConstants.CUSTOM_APPLICATION_OIDC
+                    || application?.templateId === ApplicationManagementConstants.CUSTOM_APPLICATION_SAML
                     || application?.templateId === MobileAppTemplate?.id
                     || application?.templateId === OIDCWebAppTemplate?.id
                     || application?.templateId === SinglePageAppTemplate?.id
@@ -430,7 +431,6 @@ export const applicationConfig: ApplicationConfig = {
                         render: () => (
                             <ResourceTab.Pane controlledSegmentation>
                                 <ApplicationRoles 
-                                    application={ application }
                                     onUpdate={ onApplicationUpdate }
                                 />
                             </ResourceTab.Pane>