+++ apps/console/src/features/groups/constants/group-constants.ts
@@ -49,4 +49,6 @@ export class GroupConstants {
     public static ALL_GROUPS: string = "All user stores";
 
     public static ALL_USER_STORES_OPTION_VALUE: string = "all";
+
+    public static readonly PRIMARY_USER_STORE_OPTION_VALUE: string = "primary";
 }