+++ apps/console/src/features/roles/components/application-roles.tsx
@@ -54,6 +54,7 @@ import { Grid } from "semantic-ui-react";
 import { AutoCompleteRenderOption } from "./auto-complete-render-option";
 import { ApplicationRoleWizard } from "./wizard-updated/application-role-wizard";
 import { updateApplicationDetails } from "../../applications/api";
+import { useGetApplication } from "../../applications/api/use-get-application";
 import { ApplicationInterface } from "../../applications/models";
 import {
     history
@@ -70,10 +71,6 @@ import {
 
 interface ApplicationRolesSettingsInterface extends IdentifiableComponentInterface {
     /**
-     * Application.
-     */
-    application?: ApplicationInterface
-    /**
      * on application update callback
      */
     onUpdate: (id: string) => void;
@@ -89,15 +86,18 @@ export const ApplicationRoles: FunctionComponent<ApplicationRolesSettingsInterfa
 ): ReactElement => {
 
     const {
-        application,
         onUpdate,
         [ "data-componentid" ]: componentId
     } = props;
+
+    const path: string[] = history.location.pathname.split("/");
+    const appId: string = path[path.length - 1].split("#")[0];
 
     const { t } = useTranslation();
     const dispatch: Dispatch<any> = useDispatch();
     const { getLink } = useDocumentation();
     const { organizationType } = useGetOrganizationType();
+    const { data: application } = useGetApplication(appId, !!appId);
 
     const [ isLoading, setIsLoading ] = useState<boolean>(false);
     const [ isSubmitting, setIsSubmitting ] = useState<boolean>(false);
@@ -114,9 +114,6 @@ export const ApplicationRoles: FunctionComponent<ApplicationRolesSettingsInterfa
         useState<BasicRoleInterface[]>(application?.associatedRoles?.roles ?? []);
     const [ activeOption, setActiveOption ] = useState<BasicRoleInterface>(undefined);
     const [ removedRolesOptions, setRemovedRolesOptions ] = useState<BasicRoleInterface[]>(undefined);
-
-    const path: string[] = history.location.pathname.split("/");
-    const appId: string = path[path.length - 1].split("#")[0];
 
     const isReadOnly: boolean = organizationType === OrganizationType.SUBORGANIZATION;
     const [ showWizard, setShowWizard ] = useState<boolean>(false);