+++ modules/i18n/src/models/namespaces/myaccount-ns.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ * Copyright (c) 2020, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -1026,6 +1026,7 @@ export interface MyAccountNS {
     sections: {
         accountRecovery: {
             description: string;
+            emptyPlaceholderText: string;
             heading: string;
         };
         changePassword: {